# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Roomrent API",
        "version": "1.0"
    },
    "host": "roomrent.dev",
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/register": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "create new user",
                "description": "available for new user",
                "operationId": "createUser",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "email",
                        "in": "formData",
                        "description": "email",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "username",
                        "in": "formData",
                        "description": "username",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "Name",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "description": "password",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "phone",
                        "in": "formData",
                        "description": "Phone No.",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid inputs"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "login a user",
                "description": "User must be registered",
                "operationId": "loginUser",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "identity",
                        "in": "formData",
                        "description": "username or email",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "description": "password of user",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "device_type",
                        "in": "formData",
                        "description": "type of device",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "device_token",
                        "in": "formData",
                        "description": "device specific token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid username or password"
                    }
                }
            }
        },
        "/activate/{token}": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "activates the user",
                "description": "user need to be registered",
                "operationId": "activate",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "description": "activate user with the token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "405": {
                        "description": "invalid token"
                    }
                }
            }
        },
        "/forgotpassword": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "mails for new password",
                "description": "only for already registered user",
                "operationId": "mailForgotPassword",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "email",
                        "in": "formData",
                        "description": "Email that user registered with",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "405": {
                        "description": "invalid inputs"
                    }
                }
            }
        },
        "/forgotpassword/{token}": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "check token and display form for password reset",
                "description": "user need to be registered",
                "operationId": "tokenCheckForgotPassword",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "description": "check the token for password reset",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "405": {
                        "description": "invalid token"
                    }
                }
            }
        },
        "/forgotpassword/change": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "change password of user",
                "description": "only registered user can change the password",
                "operationId": "forgotPasswordChange",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "email",
                        "in": "formData",
                        "description": "email",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "newPassword",
                        "in": "formData",
                        "description": "new password to be set",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "newPassword_confirmation",
                        "in": "formData",
                        "description": "user whose password to be changed",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "405": {
                        "description": "invalid inputs"
                    }
                }
            }
        },
        "/changepassword": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "change password of loggedin user",
                "description": "only loggedin user can change password",
                "operationId": "changePassword",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "oldPassword",
                        "in": "formData",
                        "description": "old password",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "newPassword",
                        "in": "formData",
                        "description": "new password to be set",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "405": {
                        "description": "invalid inputs"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "logout",
                "description": "user must be logged in",
                "operationId": "logout",
                "produces": [
                    "application/json"
                ],
                "parameters": [],
                "responses": {
                    "default": {
                        "description": "successfully logout"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        }
    },
...
    "securityDefinitions": {
        "api_key": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}